# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterAuthDto" }
            }
          }
        },
        "responses": { "201": { "description": "Registration successful" } },
        "summary": "Register a new local user",
        "tags": ["auth"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginAuthDto" }
            }
          }
        },
        "responses": { "201": { "description": "Login successful" } },
        "summary": "Log in with email/phone and password",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "LnQ Ecommerce API",
    "description": "The LnQ Ecommerce API specification",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "RegisterAuthDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "User last name"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890",
            "description": "Phone number"
          },
          "email": {
            "type": "string",
            "example": "john@example.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "strongPassword123",
            "description": "Password"
          }
        },
        "required": ["firstName", "email", "password"]
      },
      "LoginAuthDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "john@example.com",
            "description": "Email or phone number"
          },
          "password": {
            "type": "string",
            "example": "strongPassword123",
            "description": "Password"
          }
        },
        "required": ["identifier", "password"]
      }
    }
  }
}